server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.4
mem0ai>=0.1.48
python-dotenv==1.0.1
psycopg[pool]>=3.2.8
langchain-nvidia-ai-endpoints>=0.1.0
langchain-core>=0.2.0
qdrant-client>=1.10.0
